+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 Development
 ===========
 - (Fill this out as you fix issues and develop your features).
+- Fix for uuidRepresentation not read when provided in URI #2741
 
 Changes in 0.27.0
 =================